% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_function.R
\name{get_seoulweather}
\alias{get_seoulweather}
\title{weather function}
\usage{
get_seoulweather(start_date, end_date, personal_key)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to get Seoul weather information.
}
\examples{
get_seoulweather()
}
\keyword{cats}
